+++ frontend/src/components/NavBottom/NavBottom.css
