+++ frontend/src/components/NavItem/NavItem.jsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
 import "./NavItem.css";
-import "../assets/css/main.css";
+import "../../assets/css/main.css";
 import React from "react";
 
 // TODO a integrer et mapper dans
