+++ frontend/src/pages/Dashboard.jsx
@@ -1,6 +1,6 @@
 import BackButton from "../components/BackButton/BackButton";
-import NavBottom from "../components/Common/NavBottom/NavBottom";
-import ProfileButton from "../components/Common/ProfileButton/ProfileButton";
+import NavBottom from "../components/NavBottom/NavBottom";
+import ProfileButton from "../components/ProfileButton/ProfileButton";
 
 export default function Dashboard() {
   return (
