+++ frontend/src/pages/Home.jsx
@@ -1,5 +1,5 @@
-import MenuList from "../components/MenuList";
-import ProfileButton from "../components/Common/ProfileButton/ProfileButton";
+import MenuList from "../components/MainMenu/MainMenu";
+import ProfileButton from "../components/ProfileButton/ProfileButton";
 
 export default function Home() {
   return (
